add: temp commit

+++ mapillary_tools/lib/processing.py
@@ -3,14 +3,13 @@ import uuid
 import os
 import json
 import time
-
+import sys
 from exif_read import ExifRead
 from exif_write import ExifEdit
 from exif_aux import verify_exif
 from geo import normalize_bearing
 import config
 import uploader
-LOCAL_CONFIG_FILEPATH = uploader.LOCAL_CONFIG_FILEPATH
 STATUS_PAIRS = {"success": "failed",
                 "failed": "success"
                 }
@@ -75,7 +74,22 @@ def preform_process(import_path, root, file, process, rerun):
     return preform
 
 
-def create_and_log_process(image, import_path, mapillary_description, process, verbose=False):
+def create_and_log_process_in_list(process_file_list,
+                                   import_path,
+                                   process,
+                                   status,
+                                   verbose,
+                                   mapillary_description={}):
+    for image in process_file_list:
+        create_and_log_process(image,
+                               import_path,
+                               process,
+                               status,
+                               mapillary_description,
+                               verbose)
+
+
+def create_and_log_process(image, import_path, process, status, mapillary_description={}, verbose=False):
     # set log path
     log_root = uploader.log_rootpath(import_path, image)
     # make all the dirs if not there
@@ -86,32 +100,11 @@ def create_and_log_process(image, import_path, mapillary_description, process, v
         log_root, process)
     log_process_succes = log_process + "_success"
     log_process_failed = log_process + "_failed"
-    # if the image description is empty, the process has failed
-    if not mapillary_description:
-        if (process != "import_meta_data_process") and (process != "user_process"):
-            print("Error, " + process + " failed for image " + image)
-            open(log_process_failed, "w").close()
-            open(log_process_failed + "_" +
-                 str(time.strftime("%Y:%m:%d_%H:%M:%S", time.gmtime())), "w").close()
-            # if there is a success log from before, remove it
-            if os.path.isfile(log_process_succes):
-                os.remove(log_process_succes)
-        elif(process == "import_meta_data_process"):
-            # if there is a success log from before, remove it
-            if os.path.isfile(log_process_succes):
-                os.remove(log_process_succes)
-            # if previously read import meta properties, remove them
-            log_MAPJson = os.path.join(log_root, process + ".json")
-            if os.path.isfile(log_MAPJson):
-                if verbose:
-                    print("Warning, no import meta data set for image " + image +
-                          "in this import, previously generated import meta data will not be used")
-                os.remove(log_MAPJson)
-    # if the image description is not empty, write it in the filesystem and
-    # log success
-    else:
-        log_MAPJson = os.path.join(
-            log_root, process + ".json")
+    log_MAPJson = os.path.join(log_root, process + ".json")
+
+    if status == "success" and not mapillary_description:
+        status = "failed"
+    elif status == "success":
         try:
             save_json(mapillary_description, log_MAPJson)
             open(log_process_succes, "w").close()
@@ -124,8 +117,69 @@ def create_and_log_process(image, import_path, mapillary_description, process, v
             # if the image description could not have been written to the
             # filesystem, log failed
             print("Error, " + process + " logging failed for image " + image)
-            open(log_process_failed, "w").close()
-            open(log_process_failed + "_" +
-                 str(time.strftime("%Y:%m:%d_%H:%M:%S", time.gmtime())), "w").close()
-            if os.path.isfile(log_process_succes):
-                os.remove(log_process_succes)
+            status = "failed"
+
+    if status == "failed":
+        open(log_process_failed, "w").close()
+        open(log_process_failed + "_" +
+             str(time.strftime("%Y:%m:%d_%H:%M:%S", time.gmtime())), "w").close()
+        # if there is a success log from before, remove it
+        if os.path.isfile(log_process_succes):
+            os.remove(log_process_succes)
+        # if there is meta data from before, remove it
+        if os.path.isfile(log_MAPJson):
+            if verbose:
+                print("Warning, {} in this run has failed, previously generated properties will be removed.".format(
+                    process))
+            os.remove(log_MAPJson)
+
+
+def process_organization(mapillary_description, organization_name, organization_key, private):
+    if organization_name or organization_key:
+        if not "user_upload_token" in mapillary_description or not "MAPSettingsUserKey" in mapillary_description:
+            print(
+                "Error, can not authenticate to validate organization import, upload token or user key missing in the config.")
+            sys.exit()
+
+    if organization_name and not organization_key:
+        try:
+            organization_key = uploader.get_organization_key(mapillary_description["MAPSettingsUserKey"],
+                                                             organization_name,
+                                                             mapillary_description["user_upload_token"])
+        except:
+            print("Error, could not obtain organization key, exiting...")
+            sys.exit()
+
+    if organization_key:
+        # validate key
+        try:
+            uploader.validate_organization_key(mapillary_description["MAPSettingsUserKey"],
+                                               organization_key,
+                                               mapillary_description["user_upload_token"])
+        except:
+            print("Error, organization key validation failed, exiting...")
+            sys.exit()
+
+        # validate privacy
+        try:
+            uploader.validate_organization_privacy(mapillary_description["MAPSettingsUserKey"],
+                                                   organization_key,
+                                                   private,
+                                                   mapillary_description["user_upload_token"])
+        except:
+            print("Error, organization privacy validation failed, exiting...")
+            sys.exit()
+    return organization_key
+
+
+def inform_processing_start(import_path, len_process_file_list, process):
+    if not len_process_file_list:
+        print("No images to run {}.".format(process))
+        print("If the images have already been processed and not yet uploaded, they can be processed again, by passing the argument --rerun")
+        sys.exit()
+
+    total_file_list = uploader.get_total_file_list(import_path)
+
+    print("Running {} for {} images, skipping {} images.".format(process,
+                                                                 len_process_file_list,
+                                                                 len(total_file_list) - len_process_file_list))
